Add files via upload
js 파일 최신화

+++ static/js/index.js
@@ -1,4 +1,13 @@
 "use strict";
+/* get Token Data */
+/* parseJwt($.cookie('mytoken')) 으로 사용하면 아이디 값이 전달된다. */
+const parseJwt = (token) => {
+    try {
+        return JSON.parse(atob(token.split('.')[1]));
+    } catch (e) {
+        return null;
+    }
+};
 
 
 function All() {
@@ -82,6 +91,7 @@ function getctype(str) {
             let rows = response['posting']
             console.log(rows)
             for (let i = 0; i < rows.length; i++) {
+                let user = rows[i]['user'];
                 let num = rows[i]['num'];
                 let title = rows[i]['title']
                 let url = rows[i]['url']
@@ -100,6 +110,9 @@ function getctype(str) {
                                          <p>${star_image}</p>
                                          <p class="mycomment">${comment}</p>
                                      </div>
+                                     <footer style="text-align: right">
+                                        <div style="margin-right: 10px;">by <span style="font-weight: bold;">${user}</span>님</div>
+                                     </footer>
                                  </div>
                                  <div class="option">
                                      <a href="#" onclick="updateCard(${num});"><span class="las la-edit"></span></a>
@@ -144,11 +157,11 @@ function open_box() {
 function close_box() {
     $('#post-box').hide()
 }
-function detail_card(num){
-    window.location.href=`/detail/${num}`
+
+function detail_card(num) {
+    window.location.href = `/detail/${num}`
 
 }
-
 
 
 function listing() {
@@ -159,9 +172,11 @@ function listing() {
         success: function (response) {
             $('#card-box').empty();
             let rows = response['posting']
+            let num;
             console.log(rows);
             for (let i = 0; i < rows.length; i++) {
-                let num = rows[i]['num'];
+                let user = rows[i]['user'];
+                num = rows[i]['num'];
                 let title = rows[i]['title']
                 let url = rows[i]['url']
                 let star = rows[i]['star']
@@ -177,6 +192,14 @@ function listing() {
                                          <p>${star_image}</p>
                                          <p class="mycomment">${comment}</p>
                                      </div>
+                                     <div class="like-icon">
+                                        <i id='likeBtn-${num}' style="display: none" class="like fa fa-heart" aria-hidden="true" onclick="javascript:like(${num}, 'dislike');"></i>
+                                        <i id='dislikeBtn-${num}'  class="dislike fa fa-heart-o" aria-hidden="true" onclick="javascript:like(${num} , 'like');"></i>                                         
+                                        <p id="getLike-${num}">0</p>
+                                    </div>
+                                     <footer style="text-align: right">
+                                        <div style="margin-right: 10px;">by <span id="user" style="font-weight: bold;">${user}</span>님</div>
+                                     </footer>
                                  </div>
                                  <div class="option">
                                      <a href="#" onclick="updateCard(${num});"><span class="las la-edit"></span></a>
@@ -185,12 +208,16 @@ function listing() {
                             </div>`;
 
                 $('#card-box').append(temp_html);
+                   getlike(num);
             }
+
         }
     })
 }
 
 function posting() {
+    let token = $.cookie('mytoken');
+    let id = parseJwt(token).id;
     let url = $('#url').val()
     let title = $('#title').val()
     let star = $('#star').val()
@@ -201,7 +228,7 @@ function posting() {
     $.ajax({
         type: 'POST',
         url: '/culture',
-        data: {url_give: url, title_give: title, star_give: star, comment_give: comment, ctype_give: ctype},
+         data: {id_give:id, url_give: url, title_give: title, star_give: star, comment_give: comment, ctype_give: ctype},
         success: function (response) {
             alert(response['msg'])
             window.location.reload();
@@ -211,6 +238,13 @@ function posting() {
 }
 
 function dropList(n) {
+     let token = $.cookie('mytoken');
+    let id = parseJwt(token).id;
+    let acc = $('#user').text();
+    if(id !== acc){
+        alert('접근할 수 없습니다.')
+        return;
+    }
     $.ajax({
         type: 'DELETE',
         url: `/culture`,
@@ -223,6 +257,47 @@ function dropList(n) {
 }
 
 function updateCard(n) {
-    window.open(`http://localhost:5000/culture/update?num=${n}`, 'new', 'scrollbars=yes,resizable=no width=640 height=560, left=0,top=0\');return false')
+
+    let token = $.cookie('mytoken');
+    let id = parseJwt(token).id;
+    let acc = $('#user').text();
+
+    if(id === acc)
+        window.open(`http://localhost:5000/culture/update?num=${n}`, 'new', 'scrollbars=yes,resizable=no width=640 height=560, left=0,top=0\');return false')
+    else{
+        alert('접근할 수 없습니다.')
+        return;
+    }
 }
 
+function like(n, status){
+    if(status == 'like'){
+        $(`#likeBtn-${n}`).css('display', 'block');
+        $(`#dislikeBtn-${n}`).css('display', 'none');
+    }else{
+        $(`#likeBtn-${n}`).css('display', 'none')
+        $(`#dislikeBtn-${n}`).css('display', 'block');
+    }
+    $.ajax({
+        type: 'POST',
+        url: '/culture/like',
+        data: {num_give: n, action_give:status},
+        success: function (response) {
+            if(response['result'] =='success'){
+                getlike(n);
+            }
+        }
+    });
+}
+
+function getlike(n){
+    $.ajax({
+        type: 'GET',
+        url: `/culture/like?num_give=${n}`,
+        data: {},
+        success: function (response) {
+            let num = parseInt(response['like']);
+            $(`#getLike-${n}`).text(num);
+        }
+    });
+}
